+++ tests/pom.xml
@@ -148,6 +148,18 @@
       <artifactId>java-dotenv</artifactId>
       <version>5.2.1</version>
     </dependency>
+      <dependency>
+          <groupId>junit</groupId>
+          <artifactId>junit</artifactId>
+          <version>4.13</version>
+          <scope>compile</scope>
+      </dependency>
+    <dependency>
+      <groupId>junit</groupId>
+      <artifactId>junit</artifactId>
+      <version>4.13</version>
+      <scope>compile</scope>
+    </dependency>
 
   </dependencies>
