+++ tests/src/main/java/ca/bc/gov/open/jagefilingapi/qa/requestbuilders/GenerateUrlRequestBuilders.java
@@ -4,20 +4,26 @@ import ca.bc.gov.open.jagefilingapi.qa.backend.generateurlpayload.GenerateUrlPay
 import ca.bc.gov.open.jagefilingapi.qa.backendutils.APIResources;
 import ca.bc.gov.open.jagefilingapi.qa.backendutils.TestUtil;
 import ca.bc.gov.open.jagefilingapi.qa.config.ReadConfig;
+import ca.bc.gov.open.jagefilingapi.qa.frontend.pages.AuthenticationPage;
+import ca.bc.gov.open.jagefilingapi.qa.frontend.pages.LandingPage;
+import ca.bc.gov.open.jagefilingapi.qa.frontend.pages.PackageConfirmationPage;
+import ca.bc.gov.open.jagefilingapi.qa.frontendutils.DriverClass;
 import ca.bc.gov.open.jagefilingapi.qa.frontendutils.JsonDataReader;
 import io.restassured.RestAssured;
 import io.restassured.path.json.JsonPath;
 import io.restassured.response.Response;
 import io.restassured.specification.RequestSpecification;
-
+import org.junit.Assert;
+import org.openqa.selenium.JavascriptExecutor;
 import java.io.File;
 import java.io.IOException;
 
 import static io.restassured.RestAssured.given;
 
-public class GenerateUrlRequestBuilders {
 
-    private  RequestSpecification request;
+public class GenerateUrlRequestBuilders extends DriverClass {
+
+    private RequestSpecification request;
     private static final String X_TRANSACTION_ID = "X-Transaction-Id";
     private static final String X_USER_ID = "X-User-Id";
     private static final String GENERATE_URL_PATH = "/generateUrl";
@@ -27,8 +33,10 @@ public class GenerateUrlRequestBuilders {
     private static final String CLIENT_ID = "client_id";
     private static final String GRANT_TYPE = "grant_type";
     private static final String CLIENT_SECRET = "client_secret";
+    private static final String BASE_PATH = "user.dir";
+    private static final String PDF_PATH = "/src/test/java/testdatasource/test-document.pdf";
 
-    private  GenerateUrlPayload payloadData;
+    private GenerateUrlPayload payloadData;
 
     public Response getBearerToken() throws IOException {
         ReadConfig readConfig = new ReadConfig();
@@ -36,8 +44,8 @@ public class GenerateUrlRequestBuilders {
         String validUserid = JsonDataReader.getCsoAccountGuid().getClientSecret();
 
         request = RestAssured.given().spec(TestUtil.submitDocumentsRequestSpecification())
-                .formParam(CLIENT_ID,"efiling-demo")
-                .formParam(GRANT_TYPE,"client_credentials" )
+                .formParam(CLIENT_ID, "efiling-demo")
+                .formParam(GRANT_TYPE, "client_credentials")
                 .formParam(CLIENT_SECRET, validUserid);
 
         return request.when().post(resourceAPI).then()
@@ -50,7 +58,7 @@ public class GenerateUrlRequestBuilders {
 
         APIResources resourceAPI = APIResources.valueOf(resourceValue);
         String validExistingCSOGuid = JsonDataReader.getCsoAccountGuid().getValidExistingCSOGuid();
-        String validUserid = JsonDataReader.getCsoAccountGuid().getValidUserid();
+        String validUserid = JsonDataReader.getCsoAccountGuid().getValidUserId();
 
         Response response = getBearerToken();
         JsonPath jsonPath = new JsonPath(response.asString());
@@ -60,9 +68,9 @@ public class GenerateUrlRequestBuilders {
         File pdfFile = new File(UPLOAD_FILE_PATH + FILE_NAME_PATH);
 
         request = RestAssured.given().auth().preemptive().oauth2(accessToken).spec(TestUtil.submitDocumentsRequestSpecification())
-                .header(X_TRANSACTION_ID,validExistingCSOGuid)
-                .header(X_USER_ID,validUserid )
-                .multiPart(FILES,pdfFile);
+                .header(X_TRANSACTION_ID, validExistingCSOGuid)
+                .header(X_USER_ID, validUserid)
+                .multiPart(FILES, pdfFile);
 
         return request.when().post(resourceAPI.getResource()).then()
                 .spec(TestUtil.validDocumentResponseSpecification())
@@ -74,7 +82,7 @@ public class GenerateUrlRequestBuilders {
 
         APIResources resourceAPI = APIResources.valueOf(resourceValue);
         String validExistingCSOGuid = JsonDataReader.getCsoAccountGuid().getValidExistingCSOGuid();
-        String validUserid = JsonDataReader.getCsoAccountGuid().getValidUserid();
+        String validUserid = JsonDataReader.getCsoAccountGuid().getValidUserId();
 
         Response response = getBearerToken();
         JsonPath jsonPath = new JsonPath(response.asString());
@@ -84,28 +92,36 @@ public class GenerateUrlRequestBuilders {
         File firstPdfFile = new File(UPLOAD_FILE_PATH + FILE_NAME_PATH);
         File secondPdfFile = new File(UPLOAD_FILE_PATH + "/test-document-2.pdf");
 
-        request = RestAssured.given().auth().preemptive().oauth2(accessToken).spec(TestUtil.submitDocumentsRequestSpecification())
-                .header(X_TRANSACTION_ID,validExistingCSOGuid)
-                .header(X_USER_ID,validUserid )
-                .multiPart(FILES,firstPdfFile)
-                .multiPart(FILES,secondPdfFile);
+        request = RestAssured.given().auth().preemptive().oauth2(accessToken)
+                .spec(TestUtil.submitDocumentsRequestSpecification())
+                .header(X_TRANSACTION_ID, validExistingCSOGuid)
+                .header(X_USER_ID, validUserid)
+                .multiPart(FILES, firstPdfFile)
+                .multiPart(FILES, secondPdfFile);
 
         return request.when().post(resourceAPI.getResource()).then()
                 .spec(TestUtil.validDocumentResponseSpecification())
                 .extract().response();
     }
 
-    public Response requestWithInvalidKeyValue(String resourceValue) throws IOException {
+    public Response requestWithIncorrectFileType(String resourceValue) throws IOException {
         payloadData = new GenerateUrlPayload();
 
         APIResources resourceAPI = APIResources.valueOf(resourceValue);
         String validExistingCSOGuid = JsonDataReader.getCsoAccountGuid().getValidExistingCSOGuid();
+        String validUserid = JsonDataReader.getCsoAccountGuid().getValidUserId();
 
-        File pngFile = new File(UPLOAD_FILE_PATH + FILE_NAME_PATH);
+        Response response = getBearerToken();
+        JsonPath jsonPath = new JsonPath(response.asString());
 
-        request = RestAssured.given().spec(TestUtil.submitDocumentsRequestSpecification())
-                .header(X_TRANSACTION_ID,validExistingCSOGuid)
-                .multiPart("file",  pngFile);
+        String accessToken = jsonPath.get("access_token");
+        File pngFile = new File(UPLOAD_FILE_PATH + "/test-image-document.png");
+
+        request = RestAssured.given().auth().preemptive().oauth2(accessToken)
+                .spec(TestUtil.submitDocumentsRequestSpecification())
+                .header(X_TRANSACTION_ID, validExistingCSOGuid)
+                .header(X_USER_ID, validUserid)
+                .multiPart("file", pngFile);
 
         return request.when().post(resourceAPI.getResource()).then()
                 .spec(TestUtil.createCsoAccountIncorrectTypeErrorResponseSpecification())
@@ -117,7 +133,7 @@ public class GenerateUrlRequestBuilders {
 
         APIResources resourceAPI = APIResources.valueOf(resourceValue);
         String nonExistingCSOGuid = JsonDataReader.getCsoAccountGuid().getNonExistingCSOGuid();
-        String validUserid = JsonDataReader.getCsoAccountGuid().getValidUserid();
+        String validUserid = JsonDataReader.getCsoAccountGuid().getValidUserId();
 
         Response response = getBearerToken();
         JsonPath jsonPath = new JsonPath(response.asString());
@@ -126,9 +142,10 @@ public class GenerateUrlRequestBuilders {
 
         File pdfFile = new File(UPLOAD_FILE_PATH + FILE_NAME_PATH);
 
-        request = RestAssured.given().auth().preemptive().oauth2(accessToken).spec(TestUtil.submitDocumentsRequestSpecification())
-                .header(X_TRANSACTION_ID,nonExistingCSOGuid)
-                .header(X_USER_ID,validUserid )
+        request = RestAssured.given().auth().preemptive().oauth2(accessToken)
+                .spec(TestUtil.submitDocumentsRequestSpecification())
+                .header(X_TRANSACTION_ID, nonExistingCSOGuid)
+                .header(X_USER_ID, validUserid)
                 .multiPart(FILES, pdfFile);
 
         return request.when().post(resourceAPI.getResource()).then()
@@ -138,12 +155,21 @@ public class GenerateUrlRequestBuilders {
 
     public Response requestWithInvalidCSOAccountGuid(String resourceValue) throws IOException {
         payloadData = new GenerateUrlPayload();
-
         APIResources resourceAPI = APIResources.valueOf(resourceValue);
         String invalidNoFilingRoleGuid = JsonDataReader.getCsoAccountGuid().getInvalidNoFilingRoleGuid();
+        String inValidUserId = JsonDataReader.getCsoAccountGuid().getInValidUserId();
 
-        request = given().spec(TestUtil.requestSpecification()).header(X_TRANSACTION_ID,invalidNoFilingRoleGuid)
-                .body(payloadData.validGenerateUrlPayload());
+        Response response = getBearerToken();
+        JsonPath jsonPath = new JsonPath(response.asString());
+
+        String accessToken = jsonPath.get("access_token");
+        File pdfFile = new File(UPLOAD_FILE_PATH + FILE_NAME_PATH);
+
+        request = given().auth().preemptive().oauth2(accessToken)
+                .spec(TestUtil.submitDocumentsRequestSpecification())
+                .header(X_TRANSACTION_ID, invalidNoFilingRoleGuid)
+                .header(X_USER_ID, inValidUserId)
+                .multiPart(FILES, pdfFile);
 
         return request.when().post(resourceAPI.getResource() + invalidNoFilingRoleGuid + GENERATE_URL_PATH)
                 .then()
@@ -173,10 +199,10 @@ public class GenerateUrlRequestBuilders {
         String validExistingCSOGuid = JsonDataReader.getCsoAccountGuid().getValidExistingCSOGuid();
 
         request = given().spec(TestUtil.requestSpecification())
-                .header(X_TRANSACTION_ID,validExistingCSOGuid)
+                .header(X_TRANSACTION_ID, validExistingCSOGuid)
                 .body(payloadData.validGenerateUrlPayload());
 
-        return request.when().post(resourceInvalid.getResource()  + validExistingCSOGuid + "/generateUrs")
+        return request.when().post(resourceInvalid.getResource() + validExistingCSOGuid + "/generateUrs")
                 .then()
                 .extract().response();
     }
@@ -188,11 +214,54 @@ public class GenerateUrlRequestBuilders {
         String validExistingCSOGuid = JsonDataReader.getCsoAccountGuid().getValidExistingCSOGuid();
 
         request = given().spec(TestUtil.requestSpecification())
-                .header(X_TRANSACTION_ID,validExistingCSOGuid)
+                .header(X_TRANSACTION_ID, validExistingCSOGuid)
                 .body(payloadData.validGenerateUrlPayload());
 
         return request.when().post(resourceValid.getResource() + "generateUrl")
                 .then()
                 .extract().response();
+    }
+
+    public String getUserJwtToken() throws IOException {
+        ReadConfig readConfig = new ReadConfig();
+
+        driverSetUp();
+        String url = readConfig.getBaseUrl();
+
+        String username = System.getProperty("BCEID_USERNAME");
+        String password = System.getProperty("BCEID_PASSWORD");
+
+        driver.get(url);
+        log.info("Landing page url is accessed successfully");
+
+        // ** Leaving this step for demo mode **
+        // authenticationPage.clickBceid();
+        AuthenticationPage authenticationPage = new AuthenticationPage(driver);
+        authenticationPage.signInWithIdir(username, password);
+        log.info("user is authenticated before reaching eFiling demo page");
+
+        // ** Leaving this step for demo mode **
+        //validExistingCSOGuid = JsonDataReader.getCsoAccountGuid().getValidExistingCSOGuid();
+        //landingPage.enterAccountGuid(validExistingCSOGuid);
+        LandingPage landingPage = new LandingPage(driver);
+        String filePath = System.getProperty(BASE_PATH) + PDF_PATH;
+        landingPage.chooseFileToUpload(filePath);
+        landingPage.enterJsonData();
+        landingPage.clickEfilePackageButton();
+        log.info("Pdf file is uploaded successfully.");
+
+        // ** Leaving this step for demo mode **
+        // authenticationPage.clickBceid();
+        authenticationPage.signInWithIdir(username, password);
+        log.info("user is authenticated in eFiling demo page.");
+
+        PackageConfirmationPage packageConfirmationPage = new PackageConfirmationPage(driver);
+        boolean continuePaymentBtnIsDisplayed = packageConfirmationPage.verifyContinuePaymentBtnIsDisplayed();
+        Assert.assertTrue(continuePaymentBtnIsDisplayed);
+
+        JavascriptExecutor js = (JavascriptExecutor) driver;
+        String userToken = js.executeScript("return window.localStorage.getItem('jwt');").toString();
+        driver.quit();
+        return userToken;
     }
 }
