+++ tests/src/test/java/testdatasource/account-data.json
@@ -3,7 +3,8 @@
     "validExistingCSOGuid": "d17ff864-9686-44dd-bbf1-5ee96dcb81cb",
     "invalidNoFilingRoleGuid": "77da92db-0791-491e-8c58-1a969e67d2fb",
     "nonExistingCSOGuid": "88da92db-0791-491e-8c58-1a969e67d2fb",
-    "validUserid": "B02F2D0F367E4575BC782A414DC9365D",
+    "validUserId": "B02F2D0F367E4575BC782A414DC9365D",
+    "inValidUserId": "A02F2D0F367E4575BC782A414DC9369D",
     "clientSecret": "b3ea27bb-cc58-44a1-88a1-70a3fc3543fc"
   }
 ]
